Move classes to more logical package
