+++ android/src/main/java/bbct/android/common/activity/BaseballCardList.java
@@ -44,11 +44,11 @@ import java.util.List;
 import java.util.Objects;
 
 import bbct.android.common.R;
-import bbct.android.common.activity.util.BaseballCardActionModeCallback;
+import bbct.android.common.view.BaseballCardActionModeCallback;
 import bbct.android.common.database.BaseballCard;
 import bbct.android.common.database.BaseballCardDao;
 import bbct.android.common.database.BaseballCardDatabase;
-import bbct.android.common.provider.BaseballCardAdapter;
+import bbct.android.common.view.BaseballCardAdapter;
 import butterknife.BindView;
 import butterknife.ButterKnife;
